articles/20250804-fun-python-course/requirements.txt
aisetup
ipykernel
ipython
Pillow